refactor: Command 객체 controller 로 이동
