+++ src/main/java/chess/domain/piece/Piece.java
@@ -13,7 +13,6 @@ public abstract class Piece {
         this.color = color;
     }
 
-    // TODO: pawn 방식 수정 후 board 삭제
     public abstract void checkMovingRange(final Board board, final Position from, final Position to);
 
     public abstract boolean isPawn();
