+++ src/test/java/chess/domain/controller/CommandTest.java
@@ -1,11 +1,12 @@
-package chess.domain;
+package chess.domain.controller;
 
+import chess.controller.Command;
+import chess.domain.ChessGame;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.ValueSource;
 
-import static org.assertj.core.api.Assertions.assertThatCode;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
 
 class CommandTest {
